requirements.txt
APScheduler==2.1.1
argparse==1.2.1
distribute==0.6.24
httplib2==0.8
langid==1.1.4dev
numpy==1.7.1
oauth2==1.5.211
pymongo==2.5.2
python-twitter==1.0
simplejson==3.3.0
wsgiref==0.1.2
